Move SqlxSqliteErrorExt to utils.
