+++ discovery_engine_core/core/src/storage/sqlite.rs
@@ -52,29 +52,6 @@ use crate::storage::utils::SqlxPushTupleExt;
 // Sqlite bind limit
 const BIND_LIMIT: usize = 32766;
 
-trait SqlxSqliteErrorExt<V> {
-    /// Use this on the result of a sqlite query which might have failed a foreign
-    /// key constraint when an illegal document was passed in.
-    ///
-    /// For example it can be used in `.user_reacted` to handle it being
-    /// called with a random document id.
-    ///
-    /// If there is an error and it is a fk violation an appropriate error variant is
-    /// used instead of the default generic database error.
-    fn fk_violation_is_invalid_document_id(self, id: document::Id) -> Result<V, Error>;
-}
-
-impl<V> SqlxSqliteErrorExt<V> for Result<V, sqlx::Error> {
-    fn fk_violation_is_invalid_document_id(self, id: document::Id) -> Result<V, Error> {
-        if let Err(sqlx::Error::Database(db_err)) = &self {
-            if db_err.code() == Some(Cow::Borrowed("787")) {
-                return Err(Error::InvalidDocumentId(id));
-            }
-        }
-        self.map_err(Into::into)
-    }
-}
-
 #[derive(Clone)]
 pub(crate) struct SqliteStorage {
     pool: Pool<Sqlite>,
@@ -941,38 +918,6 @@ mod tests {
             feed[0].newscatcher_data.domain_rank,
             docs[0].newscatcher_data.domain_rank
         );
-    }
-
-    #[tokio::test]
-    async fn test_fk_violation_is_invalid_document() {
-        let storage = SqliteStorage::connect("sqlite::memory:").await.unwrap();
-
-        sqlx::query("CREATE TABLE Foo(x INTEGER PRIMARY KEY);")
-            .execute(&storage.pool)
-            .await
-            .unwrap();
-
-        sqlx::query("CREATE TABLE Bar(x INTEGER PRIMARY KEY REFERENCES Foo(x));")
-            .execute(&storage.pool)
-            .await
-            .unwrap();
-
-        let document_id = document::Id::new();
-
-        let res = sqlx::query("INSERT INTO Bar(x) VALUES (?);")
-            .bind(10)
-            .execute(&storage.pool)
-            .await
-            .fk_violation_is_invalid_document_id(document_id);
-
-        assert!(matches!(res, Err(Error::InvalidDocumentId(id)) if id == document_id));
-
-        let res = sqlx::query("malformed;")
-            .execute(&storage.pool)
-            .await
-            .fk_violation_is_invalid_document_id(document_id);
-
-        assert!(!matches!(res, Err(Error::InvalidDocumentId(_))));
     }
 
     #[tokio::test]
